@@ -4,33 +4,37 @@
     "homepage": "http://mink.behat.org/",
     "type": "library",
     "license": "MIT",
-
     "authors": [
         {
-            "name":      "Konstantin Kudryashov",
-            "email":     "ever.zet@gmail.com",
-            "homepage":  "http://everzet.com"
+            "name": "Konstantin Kudryashov",
+            "email": "ever.zet@gmail.com",
+            "homepage": "http://everzet.com"
         },
         {
-            "name":      "Christophe Coevoet",
-            "email":     "stof@notk.org"
+            "name": "Christophe Coevoet",
+            "email": "stof@notk.org"
+        },
+        {
+            "name": "Oleg Andreyev",
+            "email": "oleg@andreyev.lv"
         }
     ],
-
     "require": {
         "php": ">=7.2",
-        "behat/mink":  "^1.8"
+        "behat/mink": "^1.8"
     },
     "require-dev": {
-        "symfony/http-kernel": "^2.7 | ^3.2 | ^4.0 | ^5.0",
-        "phpunit/phpunit": "^8.2",
-        "symfony/phpunit-bridge": "^5.2"
+        "symfony/http-kernel": "^3.4 | ^4.4 | ^5.0 | ^5.4@dev",
+        "php-webdriver/webdriver": "^1.11",
+        "bamarni/composer-bin-plugin": "^1.4",
+        "phpunit/phpunit": "^9.5",
+        "symfony/phpunit-bridge": "^5.3"
     },
-
-    "bin": [
-        "bin/mink-test-server"
-    ],
-
+    "scripts": {
+        "bin": "echo 'bin not installed'",
+        "post-install-cmd": ["@composer bin all install --ansi"],
+        "post-update-cmd": ["@composer bin all update --ansi"]
+    },
     "autoload": {
         "psr-4": {
             "Behat\\Mink\\Tests\\Driver\\Util\\": "src/",
