@@ -6,7 +6,7 @@ use Behat\Mink\Tests\Driver\TestCase;
 
 class ScreenshotTest extends TestCase
 {
-    public function testScreenshot()
+    public function testScreenshot(): void
     {
         if (!extension_loaded('gd')) {
             $this->markTestSkipped('Testing screenshots requires the GD extension');
@@ -14,12 +14,12 @@ class ScreenshotTest extends TestCase
 
         $this->getSession()->visit($this->pathTo('/index.html'));
 
-        $screenShot = $this->getSession()->getScreenshot();
+        $screenshot = $this->getSession()->getScreenshot();
+        $this->assertNotEmpty($screenshot);
 
-        $this->assertIsString($screenShot);
-        $this->assertFalse(base64_decode($screenShot, true), 'The returned screenshot should not be base64-encoded');
+        $this->assertFalse(base64_decode($screenshot, true), 'The returned screenshot should not be base64-encoded');
 
-        $img = imagecreatefromstring($screenShot);
+        $img = imagecreatefromstring($screenshot);
 
         if (false === $img) {
             $this->fail('The screenshot should be a valid image');
