@@ -2,9 +2,12 @@
 
 namespace Behat\Mink\Tests\Driver;
 
+use Behat\Mink\Driver\DriverInterface;
+use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Exception\UnsupportedDriverActionException;
 use Behat\Mink\Mink;
 use Behat\Mink\Session;
+use Behat\Mink\Tests\Driver\Util\TestCaseInvalidStateException;
 use Behat\Mink\WebAssert;
 use Symfony\Bridge\PhpUnit\SetUpTearDownTrait;
 
@@ -87,6 +90,10 @@ abstract class TestCase extends \PHPUnit\Framework\TestCase
      */
     protected function getSession()
     {
+        if (!self::$mink) {
+            throw new TestCaseInvalidStateException('getSession was called before setUpBeforeClass');
+        }
+
         return self::$mink->getSession('sess');
     }
 
@@ -97,15 +104,19 @@ abstract class TestCase extends \PHPUnit\Framework\TestCase
      */
     protected function getAssertSession()
     {
+        if (!self::$mink) {
+            throw new TestCaseInvalidStateException('getSession was called before setUpBeforeClass');
+        }
+
         return self::$mink->assertSession('sess');
     }
 
     /**
      * @param string $id
      *
-     * @return \Behat\Mink\Element\NodeElement
+     * @return NodeElement
      */
-    protected function findById($id)
+    protected function findById($id): NodeElement
     {
         return $this->getAssertSession()->elementExists('named', array('id', $id));
     }
@@ -116,9 +127,9 @@ abstract class TestCase extends \PHPUnit\Framework\TestCase
      * This driver is not associated to a session. It is meant to be used for tests on the driver
      * implementation itself rather than test using the Mink API.
      *
-     * @return \Behat\Mink\Driver\DriverInterface
+     * @return DriverInterface
      */
-    protected function createDriver()
+    protected function createDriver(): DriverInterface
     {
         return self::getConfig()->createDriver();
     }
